Ensure querying the cursor position gave us something good
In rare circumstances (mostly spamming for fun), we can ask the terminal
for the position of the cursor rapidly enough that it reports garbage
before the coordinates, in which case we'll just ignore it.

This won't happen in practice, but even users who are just being silly
and holding their spacebar shouldn't be subjected to cryptic error messages.